project2/requirements.txt
flask<3.0
google-cloud-bigquery==2.24.1
